+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Full documentation for hipSOLVER is available at [hipsolver.readthedocs.io](http
 
 ## (Unreleased) hipSOLVER
 ### Added
+- Added functions
+  - gels
+    - hipsolverSSgels_bufferSize, hipsolverDDgels_bufferSize, hipsolverCCgels_bufferSize, hipsolverZZgels_bufferSize
+    - hipsolverSSgels, hipsolverDDgels, hipsolverCCgels, hipsolverZZgels
 - Added library version and device information to hipsolver-test output.
 
 ### Optimized